+++ 14/ReadMe.md
@@ -1,3 +1,12 @@
+
+### Вариант задания
+   - номер: **14**
+   
+### Версия
+ - [x] Main
+ - [ ] BugFix
+ - [ ] Revision
+ 
 ### Описание изменений
 #### 04.11.2020
 
@@ -8,6 +17,14 @@
 - Удален неиспользуемый параметр в первой строке, указывающий число параметров во второй строке
 - Добавлены комментарии в командную строку при вводе/выводе параметров
 - Прочие несущественные изменения (вроде удаления неиспользуемых директорий..)
+
+#### 12.12.2020
+
+- Создан Marven проект, проведена настройка зависимостей. 
+- Разделение логики задачи на два класса. (App + Helper)
+- Написан тест, проверяющий основной метод по двум кейсам.
+- Добавление проектных файлов в gitignore
+- Описание задачи и кейсы вынесены в readMe.txt
 
 *Knyazev Ivan, BPA18-01*
