+++ 14/javalab/readMe.txt
@@ -1,0 +1,26 @@
+
+Акерленд - город с домами, расположенными вдоль дороги. Мэр хочет установить 
+радиопередатчики на крышах домов города. Каждый передатчик имеет фиксированный 
+диапазон, что означает, что он может передавать сигнал в пределах этого расстояния.
+По карте Акерленда и дальности передачи определите минимальное количество 
+передатчиков, чтобы каждый дом находился в пределах досягаемости хотя бы одного 
+передатчика. 
+
+Например дома расположены х=х[1,2,3,5,9], передатчик имеет покрытие 1. Для покрытия всех
+домов передатчик нужно разместить на домах 2, 5 и 9, т.е. необходимо 3 передатчика.
+
+Пример #1:
+
+1 
+1 2 3 4 5
+Ответ: 2
+
+(на доме 2 и 4)
+
+Пример #2:
+
+2
+7 2 4 6 5 9 12 11
+Ответ: 3
+
+(на доме 4, 9 и 12)
